add: add authentication

+++ src/items/items.service.ts
@@ -1,7 +1,8 @@
+import { User } from '@/entities/user.entity'
 import { CreateItemDto } from '@/items/dto/create-item.dto'
 import { ItemsStatus } from '@/items/item-status.enum'
 import { ItemRepository } from '@/items/item.repository'
-import { Injectable, NotFoundException } from '@nestjs/common'
+import { BadRequestException, Injectable, NotFoundException } from '@nestjs/common'
 
 @Injectable()
 export class ItemsService {
@@ -21,12 +22,17 @@ export class ItemsService {
     return found
   }
 
-  async create(createItemDto: CreateItemDto) {
-    return await this.itemRepository.createItem(createItemDto)
+  async create(createItemDto: CreateItemDto, user: User) {
+    return await this.itemRepository.createItem(createItemDto, user)
   }
 
-  async updateStatus(id: string) {
+  async updateStatus(id: string, user: User) {
     const item = await this.findById(id)
+
+    if (item.userId === user.id) {
+      throw new BadRequestException('自身の商品を購入することはできません')
+    }
+
     const newItem = { ...item, status: ItemsStatus.SOLD_OUT, updatedAt: new Date().toISOString() }
 
     await this.itemRepository.save(newItem)
@@ -34,7 +40,13 @@ export class ItemsService {
     return newItem
   }
 
-  async delete(id: string) {
+  async delete(id: string, user: User) {
+    const item = await this.findById(id)
+
+    if (item.userId !== user.id) {
+      throw new BadRequestException('他人の商品を削除することはできません')
+    }
+
     await this.itemRepository.delete({ id })
   }
 }
